Add files via upload
Web filters are a requirement of the Children’s Internet Protection Act (CIPA). This act is concerned with protecting minors from harmful website content such as pornography, etc. Under CIPA schools and libraries are required to implement an Internet safety policy that addresses the safety and security of minors online. The proposed idea presents an approach to secure the use of BYOD policy for schools. Under BYOD policy the students are allowed to bring their own devices to school and it is important for school to have a system which manages the privileges given to student to access various sites. It helps school to monitor and control the activities of the students and gives teachers the ability to filter/block certain sites. Keywords: Web filtering, In-line page Scanning, Real-Time Contextual Analysis and Categorization, Data Mining Techniques (Pattern Matching).